+++ Tests/DartSassTests/TestVersions.swift
@@ -40,8 +40,8 @@ class TestVersions: DartSassTestCase {
     }
 
     func testVersionReport() throws {
-        let expectedPackage = "1.0.0-beta.10"
-        let expectedCompiler = "1.38.0"
+        let expectedPackage = "1.0.0-beta.11"
+        let expectedCompiler = "1.38.2"
         let compiler = try newCompiler()
         let version = try XCTUnwrap(compiler.compilerVersion.wait())
         XCTAssertEqual(expectedCompiler, version)